+++ lib/tableficate/filters/check_box.rb
@@ -1,0 +1,6 @@
+module Tableficate
+  module Filter
+    class CheckBox < Base
+    end
+  end
+end