+++ changelog.markdown
@@ -2,7 +2,8 @@
  * complete filter rewrite
  * replaced existing filter functions in table_for calls
  * added select based range filter
- * added radio button filters
+ * added radio button filter
+ * added check box filter
 
 ## 0.1.3
 * fixing html_safe break in 0.1.2